rptest: change CloudStorageTypeAndUrlStyle to list
This needs to be a list, not a tuple, because injected parameters must
be JSON compatible.

I had inadvertently changed this to a tuple for better typing in an
earlier commit, but set it back to list over a union of the types of
the two elements, which may require some casting in the future at the
call sites (when we further improve type safety), but that's fine.

See also [DEVPROD-3062] for the original fix.

Fixes CORE-14481.