rptest: block non-jsonable tests
If a test is non-jsonable (see previous commit for details), block it
from running by throwing an error in the @cluster decorator. We do the
check by literally checking if it roundtrips through json.

We try to add a detailed description of what is going on as well as
the details of the input and output objects which failed to rountrip
to help the user.