ct/l1/metastore: check abort source when purging
In the rapid topic creation test, it's seen that shutdown hangs with the
topic purger, and I think it is due to we see 100 topics we need to
purge, and with a timeout of 5 seconds each, it's probably pretty easy
to blow past our shutdown timeouts.

I wish our RPC system supported abort_sources.