lsm/db: add file_utils
This extracts a few helpers that are used in the LSM tree version class
to perform read and write operations in the database.

LevelDB Link:
- https://github.com/google/leveldb/blob/main/db/version_set.cc

Note that LevelDB exposes these awkwardly in version_set, where we pull
them into a seperate file to expose them for tests (which is much
cleaner IMO).