ducktape: Add tiered storage model test timequery debugging
The timequeries in the tiered storage model test sometimes fail to find
a message. This happens for the highest offset queried so far, close to
the max offset in the examples I've seen. This results in a generic
JSONDecodeError, which makes triaging the failure harder because it gets
lumped in with other failures. This change adds a specific assertion for
nonempty output, and some logging that will help investigate the issue
further.