ct: Take max applied epoch into account
1. The fencing is done based on max-seen-epoch value.
2. The max-seen-epoch value is not persisted to disk as part of the
   snapshot because it represents an in-flight request.

Normally, this is not the problem. When every placeholder batch is
applied we're updating both max-seen and max-applied epoch values
('ctp_stm_state::advance_epoch' method). The problem is that right after
the snapshot is loaded the max-seen-epoch is nullopt so it is possible
to acquire a fence for every epoch.

This commit fixes this by checking both max-seen and max-applied.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>