partition/timequery: Fix tiered storage timequery bug
This change refactors timequery logic to fix a rare bug where a
timequery would return no result despite an offset matching the
timestamp existing in cloud storage and local storage.

The bug is observed in the tiered storage model test when querying
timestamps corresponding to offsets in the final, active segment of the
log of a partition with tiered storage enabled and with aggressive
cleanup settings. The partition gets into a state where the local log is
truncated up to the high watermark, but the active segment remains.
Querying for a timestamp within the bounds of the active segment
meant that the timestamp was after the start timestamp for the local
log, which comes from the base timestamp of the only segment, but the
start offset of the log was after the max offset set by the timequery,
which is the HWM and comes from the Kafka handler. This caused the
timequery to return no result, incorrectly, without checking local or
cloud storage.

The refactor fixes the bug and hopefully cleans up the logic a bit.
Regression test included.