rpk: move cloud auth token to its own package
This will allow us to use this lightweight
package as helper in other packages w/o
circular dependencies with the bigger
package (oauth). We are expanding our usage
as we introduce more publicAPI commands.

An example usage is introduced in the
controlplane client where we are now able
to validate the token before creating the
client.