rptest/omb: update open-messaging benchmark
To 9cff468d5124cc9eb6cfce90798c2c9a5c354f6c. This pulls in nearly a year
of changes, though many are to "deploy" which don't affect runtime in
the DT context.

The driver was for the tip change (9cff468) which adds logging which we
want to diagnose possible hangs in the serverless test.