rptest/ruff: enable ruff check in CI
Enable the ruff linter in CI. It has many useful checks and
rptest is currently clean except for the two exceptions in .ruff.toml.

We restrict it to tests subdir currently as outside of tests there are
still some unclean files, especially in tools.

We use --output-format=github in the GitHub Action so that the errors
are shown inline in the PR view (as "annotations").