rptest/consumer_group_test: more cleanup on fail
This test has been observing hanging after failure: tear down completes
but the process does not finish. This happens when the test itself is
failing to do an exception. See for example test_large_group_count in:

https://ci-artifacts.dev.vectorized.cloud/production-devprod/vtools/26190/019ab4e2-0265-4063-8faa-0c08b34b9266/vbuild/ducktape/results/2025-11-24--001/report.html

The most likely candidates is threads hanging around which have not
exited. This is a corrupting failure, i.e., can fail other
concurrently running tests.

We are making some changes on the DT side to help diagnose these better
but for this specific test, do the following:

 - Always close the event loop, even on expcetion, using context manager
 - In asycio.gather execute all the tasks if some fail, so that they
   have time to finish and clean up
 - In the group poll tasks, use try/finally to ensure we close the
   consumer even on exception

[CORE-13453]