k/client: Add option to include authorized operations
Not all uses of the kafka client need metadata or describe_cluster
requests to include the bitmask of authorized operations.  If the
authenticated principal used by the client doesn't have the appropriate
level of authorization to get that information this clutters up the
broker logs with failed authorization checks.  This new option permits
users of k/client to opt out of getting that information.

Signed-off-by: Michael Boquard <michael@redpanda.com>