dt: Longify all scram passwords.
For fips 140-3, hmac keys need to be at least 112 bits.
The first operation in scram is to compute an hmac with the password as
the key. So, all scram passwords need to be at least 14 bytes long, in
fips. Update all passwords in dt to be at least 14 bytes long.