`ct/l1`: add `file_arena`
To be used as a scratch space manager for L1 objects in `cloud_topics`.

A very faithful copy-paste/rework of the `datalake` manager- the work stealing
approach with shard-local pools and a global memory reservation is reproduced
here. The same approach of a soft-limit and a manager on core0 which
reclaims unused reservations from each shard is utilized.

The `file_arena` acts as a wrapper around the `cloud_topics::l1::io` layer,
in which users request `staging_file`s with their expected maximum file size
up front.

The user is returned a `staging_file_with_reservation` object, which is also a
thin wrapper, this time around a `staging_file` and some semaphore units.
The `staging_file` can be utilized as expected, and after the user is finished
writing to it, can call `finalize()` in order to return any unused space
(the delta between the file size on disk and the up front memory reservation)
to the shard-local pool. Finally, when the file is no longer required, the
`remove()` function invokes `staging_file::remove()` and also returns all
semaphore units back to the shard-local pool. This two-stage return approach
is helpful for `cloud_topics`- both the `reconciler` and compaction use L1
files of an target maximum size, but have an interdeterminate amount of time
between finalizing their writes to the file (which may end up being smaller than
the expected maximum) and uploading/removing the files from disk.

Obviously, this means a `file_arena` must outlive all `staging_file_with_reservation`
objects.