base_transport: use optional<input_stream>
Change the input_stream member of base_transport to be an
optional<input_stream>, constructed in do_connect and destroyed in
stop().

Additionally, close() input stream on stop():

`close()` the input_stream and reset the optional, destroying the stream
in `base_transport::stop()`. This is both good hygiene and also aligns
with upstream where the assignment and default constructor of
input_stream are deprecated: in particular, assignment apparently does
not necessarily do the right thing if the stream has ever been used.

Finally, we also hide the _in member and use the in() accessor to get to
it, so we can assert the optional has a value, and easy further
implementation changes.