seastar-26.1: update generator usages
The generator implementation in upstream seastar was re-written, and
one implication is that `co_await gen()` where gen() is a generator
expression now returns optional<reference_wrapper<T>> instead of
optional<T>. So we need an extra `get()` in there somewhere to
deference the reference wrapper.

We had a few generator uses in Redpanda, this change updates them per
above.