`ct/l1`: add `metastore::get_extent_metadata_{forwards/backwards}`
Returns a number of extents in the offset range `[start, end]`
inclusively, and in ascending or descending offset order (for `forwards()`/
`backwards()` respectively).

`forwards()` is obviously useful for forward iteration over an extent-aligned
offset range- that is, for an extent metastore state of `[[0, 9],[10,19],[20,29]]`
and a request like `get_extent_metadata_forwards([0, 15])`, the returned
extents will be `[[0, 9], [10, 19]]`.

`backwards()` is obviously useful for backward iteration over an extent-aligned
offset range- that is, for an extent metastore state of `[[0, 9],[10,19],[20,29]]`
and a request like `get_extent_metadata_backwards([0, 15])`, the returned extents
will be `[[10, 19], [0, 9]]`.