ct: Replace map with atmoic counters
Currently, the write_request_scheduler uses 'map' method to collect
information from every shard. This is done on shard zero. The tasks are
submitted to every shard's queue and then awaited. This operation has
relatively high latency.

This commit changes this algorithm a bit. In the start method the
scheduler collects references to counters of all stages. The counters
are atomic. Instead of using map the time-based-fallback policy just
collects counter values directly.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>