k/group_mgr: use quorum_ack replication for group tombstone replication
When group tombstone is replicated previously we were using `leader_ack`
consistency. It is optimization that is not really relevant in this case
as leader acks are there to increase the throughput by minimizing the
latency of request handling.

Another point is that it is very likely the tombstone replicate request
will be immediately followed by `offset_commit` request that will fore
flush either way.

Changed the consistency level used to replicate the tombstones with
`quorum_acks`

Signed-off-by: Michał Maślanka <michal@redpanda.com>