`ct/l1`: `remove()/close()` unused inflight objects
We might reach the state where `_inflight_object` _was_ initialized
because there was compaction work to be done, but due to some `metastore`
RPC error or some pre-emption from shutdown we reach `finalize()` without
any extents processed.

In this case, `remove()` and `close()` the unused inflight objects,
and proceed to finalize the job as normal.